refactor(#5): split watchlist manage client into hooks and panels
Extract company/stakeholder hooks, filters, and tab panels so MarketSignalsManageClient stays a thin orchestrator.
