+++ app/dashboard/settings/market-signals/watchlist-manage-types.ts
@@ -1,0 +1,18 @@
+export type ManageCompany = {
+  id: string
+  name: string
+  logoUrl: string | null
+  isFollowing: boolean
+  accountStatus: string | null
+}
+
+export type WatchedStakeholder = {
+  key: string
+  personName: string
+  companyName: string | null
+  personTitle: string | null
+  createdAt: string
+  isFollowing: boolean
+}
+
+export type ManageWatchlistTab = 'companies' | 'executives'
